+++ next-app/src/components/common/Card/FeedItem.style.tsx
