+++ next-app/src/components/common/Card/GridType.tsx
@@ -2,7 +2,7 @@ import Image from 'next/image'
 
 import Icons from 'assets/icons'
 import { colors } from 'styles/colors'
-import * as S from './Card.style'
+import * as S from './FeedItem.style'
 import { Container, GridTypeWrapper } from './GridType.style'
 
 const GridType = () => {
