+++ next-app/src/components/common/Card/index.ts
@@ -1,1 +1,1 @@
-export { default } from './Card'
+export { default } from './FeedItem'
